pull default ic
